Modelling loading, unit send-to-location AI with coordinates drawn by fragment shader at click point, beginning AI and abstracting units

+++ PowerTownSim/Surface.h
@@ -25,7 +25,7 @@ private:
 	unsigned int vao;
 
 	std::vector <float> vertices;
-	std::vector <GLshort> indices;
+	std::vector <GLuint> indices;
 
 	// Geometric variables
